13_webinar_python_django/requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
whitenoise
